Fixed type traits to support derived types
As it was the type traits interface had a specialization for std::vector and a generic template for all other types. This caused issues for types that derive from vector, as they matched the generic traits and not the specialized one. I have removed the specialization logic so that the traits simply try to get the interfaces if they exist. This was all necessary because i wanted to use a type that privately inherited from std::vector